RAG_Chatbot/requirements.txt
streamlit==1.38.0
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
ollama
langchain-ollama